backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.39.0
huggingface-hub>=0.22.0
pydantic>=2.6.0
python-dotenv>=1.0.1
nltk>=3.8.1
pillow>=10.2.0
requests>=2.31.0
numpy>=1.24.4
scipy>=1.10.1
opencv-python>=4.9.0.80
python-multipart>=0.0.6